+++ index.html
@@ -167,7 +167,7 @@
   </template>
 
   <!-- Подключим скрипт -->
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 
 </html>
